AdvancedDrawContext: Move `DrawContext.drawTexture(UGpuTextureView)` to separate file
Because the `AdvancedDrawContext` itself will cease to exist with 26.3,
but drawing a texture to the vanilla draw context is still very much
needed.
